+++ src/main/java/org/openjdk/shenandoah/DataConnector.java
@@ -1,6 +1,6 @@
 /*
- * Copyright (c) 2021, Red Hat, Inc. All rights reserved.
- * Copyright (c) 2021, Amazon.com, Inc. All rights reserved.
+ * Copyright (c) 2023, Red Hat, Inc. All rights reserved.
+ * Copyright (c) 2023, Amazon.com, Inc. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
