+++ src/main/java/org/openjdk/shenandoah/RegionAffiliation.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020, Amazon.com, Inc. or its affiliates All rights reserved.
+ * Copyright (c) 2023, Amazon.com, Inc. or its affiliates All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
